ES 时间格式化

+++ src/main/java/com/example/demo/dao/rabc/AuthoritiesMapper.java
@@ -35,5 +35,5 @@ public interface AuthoritiesMapper {
 
     List<AuthoritiesVM> getPageData(AuthoritiesVM viewModel);
 
-    Integer getPageDataCount(AuthoritiesVM viewModel);
+    Long getPageDataCount(AuthoritiesVM viewModel);
 }
