+++ src/main/java/com/example/demo/dao/rabc/MenusMapper.java
@@ -35,5 +35,5 @@ public interface MenusMapper {
 
     List<MenusVM> getPageData(MenusVM viewModel);
 
-    Integer getPageDataCount(MenusVM viewModel);
+    Long getPageDataCount(MenusVM viewModel);
 }
