+++ src/main/java/com/example/demo/service/rabc/RoleMenuAuthoritiesService.java
@@ -84,7 +84,7 @@ public class RoleMenuAuthoritiesService {
         MessageResult<PageData<RoleMenuAuthoritiesVM>> message = new MessageResult<>();
         try {
             PageData<RoleMenuAuthoritiesVM> paegData = new PageData<>();
-            Integer count = roleMenuAuthoritiesMapper.getPageDataCount(viewModel);
+            Long count = roleMenuAuthoritiesMapper.getPageDataCount(viewModel);
             paegData.setCount(count);
             List<RoleMenuAuthoritiesVM> list = roleMenuAuthoritiesMapper.getPageData(viewModel);
             paegData.setData(list);
